Fix type genericity in constraint Jacobian allocations
Replace hardcoded Float64 with eltype(θ) in constraint Jacobian
allocation patterns to properly support BigFloat and other numeric types.

Changes:
- OptimizationLBFGSB: Use eltype(θ) instead of Float64 in aug_grad
- OptimizationAuglag: Use eltype(θ) instead of Float64 in aug_grad
- OptimizationEnzymeExt: Remove Float64 return type annotation from lagrangian
- Add interface compatibility tests for BigFloat support

+++ lib/OptimizationBase/ext/OptimizationEnzymeExt.jl
@@ -66,7 +66,7 @@ function cons_f2_oop(mode, x, dx, fcons, p, i)
     return nothing
 end
 
-function lagrangian(x, _f::Function, cons::Function, p, λ, σ = one(eltype(x)))::Float64
+function lagrangian(x, _f::Function, cons::Function, p, λ, σ = one(eltype(x)))
     res = zeros(eltype(x), length(λ))
     cons(res, x, p)
     return σ * _f(x, p) + dot(λ, res)
